refactor(proxy): move bridge activity lifecycle helpers
